requirements.txt
bs4==0.0.2
SQLAlchemy==2.0.30
python-dotenv==1.0.0
fastapi==0.111.0
